splited logic to hooks from alerts counter component
